requirements.txt
ldap3==2.9.1
prometheus-client==0.15.0
pyasn1==0.4.8
PyYAML==6.0
